+++ src/main/java/com/buuz135/simpleclaims/claim/ClaimManager.java
@@ -16,6 +16,7 @@ import com.hypixel.hytale.logger.HytaleLogger;
 import com.hypixel.hytale.math.util.ChunkUtil;
 import com.hypixel.hytale.server.core.entity.entities.Player;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
+import com.hypixel.hytale.server.core.universe.Universe;
 import com.hypixel.hytale.server.core.universe.world.World;
 import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import it.unimi.dsi.fastutil.longs.LongSet;
@@ -178,7 +179,7 @@ public class ClaimManager {
         if (adminOverridesBlockingFile.getAdminOverrides().contains(playerUUID)) return true;
 
         var chunkInfo = getChunkRawCoords(dimension, chunkX, chunkZ);
-        if (chunkInfo == null) return true;
+        if (chunkInfo == null) return !Arrays.asList(Main.CONFIG.get().getFullWorldProtection()).contains(dimension);
 
         var chunkParty = getPartyById(chunkInfo.getPartyOwner());
         if (chunkParty == null || interactMethod.test(chunkParty)) return true;
@@ -352,7 +353,7 @@ public class ClaimManager {
 
     public void queueMapUpdateForParty(PartyInfo partyInfo) {
         this.getChunks().forEach((dimension, chunkInfos) -> {
-            var world = Main.WORLDS.get(dimension);
+            var world = Universe.get().getWorlds().get(dimension);
             if (world != null) {
                 for (ChunkInfo value : chunkInfos.values()) {
                     if (value.getPartyOwner().equals(partyInfo.getId())) {