+++ src/main/java/com/buuz135/simpleclaims/config/SimpleClaimsConfig.java
@@ -43,6 +43,12 @@ public class SimpleClaimsConfig {
             .append(new KeyedCodec<String[]>("WorldClaimBlacklist", Codec.STRING_ARRAY),
                     (simpleClaimsConfig, value, extraInfo) -> simpleClaimsConfig.WorldNameBlacklistForClaiming = value,
                     (simpleClaimsConfig, extraInfo) -> simpleClaimsConfig.WorldNameBlacklistForClaiming).add()
+            .append(new KeyedCodec<String>("TitleTopClaimTitleText", Codec.STRING),
+                    (simpleClaimsConfig, value, extraInfo) -> simpleClaimsConfig.TitleTopClaimTitleText = value,
+                    (simpleClaimsConfig, extraInfo) -> simpleClaimsConfig.TitleTopClaimTitleText).add()
+            .append(new KeyedCodec<String[]>("FullWorldProtection", Codec.STRING_ARRAY),
+                    (simpleClaimsConfig, value, extraInfo) -> simpleClaimsConfig.FullWorldProtection = value,
+                    (simpleClaimsConfig, extraInfo) -> simpleClaimsConfig.FullWorldProtection).add()
             .build();
 
     private int DefaultPartyClaimsAmount = 25;
@@ -55,6 +61,8 @@ public class SimpleClaimsConfig {
     private boolean AllowPartyBreakBlockSetting = true;
     private boolean AllowPartyInteractBlockSetting = true;
     private String[] WorldNameBlacklistForClaiming = new String[0];
+    private String TitleTopClaimTitleText = "Simple Claims";
+    private String[] FullWorldProtection = new String[0];
 
     private boolean ForceSimpleClaimsChunkWorldMap = true;
     private boolean CreativeModeBypassProtection = false;
@@ -109,5 +117,13 @@ public class SimpleClaimsConfig {
 
     public boolean isAllowPartyInteractBlockSetting() {
         return AllowPartyInteractBlockSetting;
+    }
+
+    public String getTitleTopClaimTitleText() {
+        return TitleTopClaimTitleText;
+    }
+
+    public String[] getFullWorldProtection() {
+        return FullWorldProtection;
     }
 }