+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.0.16
+
+* Added a claimed chunks list for the party, closes #50
+* Added a config option to change the title top line when entering a claim, closes #16
+* Fixed a bug where the admin-modify-chunk-all command wasn't updating parties amount if they had the same amount as the
+  config
+
+# 1.0.14
+
+* Fixed inviting players to a party using the GUI not working properly
+* Changed containers to use decorated containers
+
 # 1.0.13
 
 * Filter out players that have a party from the invite dropdown